+++ src/models/general_structs/mod.rs
@@ -1,0 +1,4 @@
+pub mod members;
+pub mod metadata;
+pub mod privacy;
+pub mod references;